[03136] Extract shared plan creation dialog logic into CreatePlanDialogLauncher
